Module cleanup
Started cleanin up the program (to make it faster) including removing all the unused modules. I also changed the timeouts in the scanning class in Open_Cliker.pyw to wait 100ms when the program is not on ans 1ms when it is on. the timeout change was added to significantly lower the cpu usage of the proogram. Last the wierd file inconsistencys that broke clear_tmp_data.pyw have been fixed